+++ CHANGES.md
@@ -5,6 +5,8 @@
 * Added `selected-any?` operation that returns true if any element is navigated to.
 * Huge performance improvements to `select`, `select-one`, `select-first`, and `select-one!`
 * Added META navigator (thanks @aengelberg)
+* Added DISPENSE navigator to drop all collected values for subsequent navigation
+* Added `collected?` macro to create a filter function which operates on the collected values
 * Bug fix: END, BEGINNING, FIRST, and LAST now work properly on nil
 
 ## 0.11.2